Split cache into runtime specific code
