+++ testkit/jvm/src/main/scala/cats/effect/testkit/TestInstancesPlatform.scala
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2020-2022 Typelevel
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cats.effect
+package testkit
+
+import java.{util => ju}
+
+trait TestInstancesPlatform { self: TestInstances =>
+  private[this] val tickerRuntimeCache: ju.Map[Ticker, unsafe.IORuntime] =
+    ju.Collections.synchronizedMap(new ju.WeakHashMap())
+
+  implicit def materializeRuntime(implicit ticker: Ticker): unsafe.IORuntime = {
+    val runtime = tickerRuntimeCache.get(ticker)
+    if (runtime ne null) {
+      runtime
+    } else {
+      unsafe.IORuntime.testRuntime(ticker.ctx, scheduler)
+    }
+  }
+}
